requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Google Gemini
google-generativeai==0.3.2

# Web scraping
selenium==4.17.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.3

# Data processing
numpy==1.26.3
scikit-learn==1.4.0

# HTTP requests
requests==2.31.0

# Development tools
python-dotenv==1.0.0
